update : cent system integrated

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('balance')->default(0);
+            $table->bigInteger('balance')->default(0); // Stored in cents
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
